Apply 8-bit conversion and LUT to result image.
When applying a classifier from macro, make sure the result is in 8-bit and with the same lookup table as in the GUI.

TODO: use the same method from the GUI and the macro. As it is now, the GUI accept a list of images but the macro command processes only one.