Fix bug when opening large images.
As reported in the ImageJ forum (https://forum.image.sc/t/trainable-weka-segmentation-issues-suggestions/22588/3) the GUI got larger than the screen dimensions when large images were used and it could not be resized. Now the problem is fixed by zooming out if any of the image canvas dimensions are larger than 0.75 of the screen dimensions.