Fix bug created on commit #4368a7f.
Basically features need to be updated after loading a new classifier to avoid state conflits in the plugin (e.g., features set as updated in training but not available after the classifier loading).

To prevent that, we call now setFeaturesDirty() in adjustSegmentationStateToData() if features have changed.

Also, there is no need to check if the feature stacks are empty in setFeaturesDirty(): all slices with traces will be set to be updated during training and the rest during test.